fix more things
