+++ ext/ManoptManifoldsExt/ManoptManifoldsExt.jl
@@ -4,9 +4,17 @@ if isdefined(Base, :get_extension)
     using Manifolds
     using ManifoldsBase
     using Manopt
-    import Manopt: max_stepsize, prox_TV2, grad_TV2
+    import Manopt:
+        max_stepsize,
+        prox_TV2,
+        grad_TV2,
+        alternating_gradient_descent,
+        alternating_gradient_descent!,
+        get_gradient,
+        get_gradient!
     using StaticArrays
-    using LinearAlgebra: cholesky, det, diag, dot, Hermitian, qr, Symmetric, triu
+    using LinearAlgebra:
+        cholesky, det, diag, dot, Hermitian, qr, Symmetric, triu, I, Diagonal
     import ManifoldsBase: copy, mid_point, mid_point!
 
     using ManifoldDiff:
@@ -18,9 +26,17 @@ else
     using ..Manifolds
     using ..ManifoldsBase
     using ..Manopt
-    import ..Manopt: max_stepsize, prox_TV2, grad_TV2
+    import ..Manopt:
+        max_stepsize,
+        prox_TV2,
+        grad_TV2,
+        alternating_gradient_descent,
+        alternating_gradient_descent!,
+        get_gradient,
+        get_gradient!
     using ..StaticArrays
-    using ..LinearAlgebra: cholesky, det, diag, dot, Hermitian, qr, Symmetric, triu
+    using ..LinearAlgebra:
+        cholesky, det, diag, dot, Hermitian, qr, Symmetric, triu, I, Diagonal
     import ..ManifoldsBase: copy, mid_point, mid_point!
 
     using ..ManifoldDiff:
@@ -33,5 +49,6 @@ include("manifold_functions.jl")
 include("nonmutating_manifolds_functions.jl")
 include("artificialDataFunctionsManifolds.jl")
 include("ChambollePockManifolds.jl")
+include("alternating_gradient.jl")
 
 end
