+++ ext/ManoptManifoldsExt/artificialDataFunctionsManifolds.jl
@@ -1,42 +1,4 @@
 
-@doc raw"""
-    artificial_S2_composite_bezier_curve()
-
-Create the artificial curve in the `Sphere(2)` consisting of 3 segments between the four
-points
-
-````math
-p_0 = \begin{bmatrix}0&0&1\end{bmatrix}^{\mathrm{T}},
-p_1 = \begin{bmatrix}0&-1&0\end{bmatrix}^{\mathrm{T}},
-p_2 = \begin{bmatrix}-1&0&0\end{bmatrix}^{\mathrm{T}},
-p_3 = \begin{bmatrix}0&0&-1\end{bmatrix}^{\mathrm{T}},
-````
-
-where each segment is a cubic Bezér curve, i.e. each point, except $p_3$ has a first point
-within the following segment $b_i^+$, $i=0,1,2$ and a last point within the previous
-segment, except for $p_0$, which are denoted by $b_i^-$, $i=1,2,3$.
-This curve is differentiable by the conditions $b_i^- = \gamma_{b_i^+,p_i}(2)$, $i=1,2$,
-where $\gamma_{a,b}$ is the [`shortest_geodesic`](https://juliamanifolds.github.io/Manifolds.jl/stable/interface.html#ManifoldsBase.shortest_geodesic-Tuple{Manifold,Any,Any}) connecting $a$ and $b$.
-The remaining points are defined as
-
-````math
-\begin{aligned}
-    b_0^+ &= \exp_{p_0}\frac{\pi}{8\sqrt{2}}\begin{pmatrix}1&-1&0\end{pmatrix}^{\mathrm{T}},&
-    b_1^+ &= \exp_{p_1}-\frac{\pi}{4\sqrt{2}}\begin{pmatrix}-1&0&1\end{pmatrix}^{\mathrm{T}},\\
-    b_2^+ &= \exp_{p_2}\frac{\pi}{4\sqrt{2}}\begin{pmatrix}0&1&-1\end{pmatrix}^{\mathrm{T}},&
-    b_3^- &= \exp_{p_3}-\frac{\pi}{8\sqrt{2}}\begin{pmatrix}-1&1&0\end{pmatrix}^{\mathrm{T}}.
-\end{aligned}
-````
-
-This example was used within minimization of acceleration of the paper
-
-> Bergmann, R., Gousenbourger, P.-Y.:
-> _A variational model for data fitting on manifolds by minimizing the acceleration of a Bézier curve_,
-> Front. Appl. Math. Stat. 12, 2018.
-> doi: [10.3389/fams.2018.00059](https://dx.doi.org/10.3389/fams.2018.00059)
-> arxiv: [1807.10090](https://arxiv.org/abs/1807.10090)
-
-"""
 function Manopt.artificial_S2_composite_bezier_curve()
     M = Sphere(2)
     d0 = [0.0, 0.0, 1.0]
@@ -75,24 +37,6 @@ function Manopt.artificial_S2_composite_bezier_curve()
     ]
 end
 
-@doc raw"""
-    artificial_S2_rotation_image([pts=64, rotations=(.5,.5)])
-
-Create an image with a rotation on each axis as a parametrization.
-
-# Optional Parameters
-* `pts` – (`64`) number of pixels along one dimension
-* `rotations` – (`(.5,.5)`) number of total rotations performed on the axes.
-
-This dataset was used in the numerical example of Section 5.1 of
-
-> Bačák, M., Bergmann, R., Steidl, G., Weinmann, A.:
-> _A Second Order Non-Smooth Variational Model for Restoring Manifold-Valued Images_
-> SIAM J. Sci. Comput. 38(1), A567–A597, 2016.
-> doi: [10.1137/15M101988X](https://dx.doi.org/10.1137/15M101988X)
-> arxiv: [1506.02409](https://arxiv.org/abs/1506.02409)
-
-"""
 function Manopt.artificial_S2_rotation_image(
     pts::Int=64, rotations::Tuple{Float64,Float64}=(0.5, 0.5)
 )
@@ -111,31 +55,6 @@ function Manopt.artificial_S2_rotation_image(
     return img
 end
 
-@doc raw"""
-    artificial_S2_lemniscate(p, t::Float64; a::Float64=π/2)
-
-Generate a point from the signal on the [`Sphere`](https://juliamanifolds.github.io/Manifolds.jl/stable/manifolds/sphere.html) $\mathbb S^2$ by
-creating the [Lemniscate of Bernoulli](https://en.wikipedia.org/wiki/Lemniscate_of_Bernoulli)
-in the tangent space of `p` sampled at `t` and use èxp` to obtain a point on
-the [`Sphere`](https://juliamanifolds.github.io/Manifolds.jl/stable/manifolds/sphere.html).
-
-# Input
-* `p` – the tangent space the Lemniscate is created in
-* `t` – value to sample the Lemniscate at
-
-# Optional Values
- * `a` – (`π/2`) defines a half axis of the Lemniscate to cover a
-   half sphere.
-
-This dataset was used in the numerical example of Section 5.1 of
-
-> Bačák, M., Bergmann, R., Steidl, G., Weinmann, A.:
-> _A Second Order Non-Smooth Variational Model for Restoring Manifold-Valued Images_
-> SIAM J. Sci. Comput. 38(1), A567–A597, 2016.
-> doi: [10.1137/15M101988X](https://dx.doi.org/10.1137/15M101988X)
-> arxiv: [1506.02409](https://arxiv.org/abs/1506.02409)
-
- """
 function Manopt.artificial_S2_lemniscate(p, t::Float64, a::Float64=π / 2.0)
     M = Sphere(2)
     tP = 2.0 * Float64(p[1] >= 0.0) - 1.0 # Take north or south pole
@@ -146,24 +65,6 @@ function Manopt.artificial_S2_lemniscate(p, t::Float64, a::Float64=π / 2.0)
     return exp(M, p, tV)
 end
 
-@doc raw"""
-    artificial_S2_whirl_image([pts=64])
-
-Generate an artificial image of data on the 2 sphere,
-
-# Arguments
-* `pts` – (`64`) size of the image in `pts`$\times$`pts` pixel.
-
-This example dataset was used in the numerical example in Section 5.5 of
-
-> Laus, F., Nikolova, M., Persch, J., Steidl, G.:
-> _A Nonlocal Denoising Algorithm for Manifold-Valued Images Using Second Order Statistics_,
-> SIAM J. Imaging Sci., 10(1), 416–448, 2017.
-> doi: [ 10.1137/16M1087114](https://dx.doi.org/10.1137/16M1087114)
-> arxiv: [1607.08481](https://arxiv.org/abs/1607.08481)
-
-It is based on [`artificial_S2_rotation_image`](@ref) extended by small whirl patches.
-"""
 function Manopt.artificial_S2_whirl_image(pts::Int=64)
     M = Sphere(2)
     img = artificial_S2_rotation_image(pts, (0.5, 0.5))
@@ -195,4 +96,48 @@ function Manopt.artificial_S2_whirl_image(pts::Int=64)
         img[r...] = patch
     end
     return img
+end
+
+function Manopt.artificial_SPD_image2(pts=64, fraction=0.66)
+    Zl = 4.0 * Matrix{Float64}(I, 3, 3)
+    # create a first matrix
+    α = 2.0 * π / 3
+    β = π / 3
+    B = [1.0 0.0 0.0; 0.0 cos(β) -sin(β); 0.0 sin(β) cos(β)]
+    A = [cos(α) -sin(α) 0.0; sin(α) cos(α) 0.0; 0.0 0.0 1.0]
+    Zo = Matrix(Symmetric(A * B * Diagonal([2.0, 4.0, 8.0]) * B' * A'))
+    # create a second matrix
+    α = -4.0 * π / 3
+    β = -π / 3
+    B = [1.0 0.0 0.0; 0.0 cos(β) -sin(β); 0.0 sin(β) cos(β)]
+    A = [cos(α) -sin(α) 0.0; sin(α) cos(α) 0.0; 0.0 0.0 1.0]
+    Zt = A * B * Diagonal([8.0 / sqrt(2.0), 8.0, sqrt(2.0)]) * B' * A'
+    data = fill(Matrix{Float64}(I, 3, 3), pts, pts)
+    M = SymmetricPositiveDefinite(3)
+    for row in 1:pts
+        for col in 1:pts
+            # (a) from Zo a part to Zt
+            C = Zo
+            if (row > 1) # in X direction
+                C = exp(
+                    M,
+                    C,
+                    log(M, C, Zt),
+                    (row - 1) / (2 * (pts - 1)) + ((row > fraction * pts) ? 1 / 2 : 0.0),
+                )
+            end
+            if (col > 1) # and then in Y direction
+                C = exp(
+                    M,
+                    C,
+                    vector_transport_to(
+                        M, Symmetric(Zo), log(M, Zo, Zl), Symmetric(C), ParallelTransport()
+                    ),
+                    (col - 1.0) / (pts - 1),
+                )
+            end
+            data[row, col] = C
+        end
+    end
+    return data
 end
